backend/requirement.txt
attrs==20.2.0
boto3==1.15.11
botocore==1.18.11
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.3
django-cors-headers==2.4.0
django-filter==2.2.0
django-jsonfield==1.4.0
django-templated-mail==1.1.1
django-user-agents==0.4.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
loguru==0.4.1
Markdown==3.0.1
MarkupSafe==1.1.1
mysqlclient==1.3.13
packaging==20.4
Pillow==7.2.0
PyJWT==1.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
six==1.15.0
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.10
user-agents==2.2.0
